Create folders for each Necessity Component
