+++ orange/src/components/Necessity/NecessityHead/NecessityHead.css
@@ -1,0 +1,26 @@
+.NecessityHeadBlock {
+    padding-top: 48px;
+    padding-left: 32px;
+    padding-right: 32px;
+    padding-bottom: 24px;
+    border-bottom: 1px solid #e9ecef;
+}
+
+.NecessityHeadBlock h1 {
+    margin: 0;
+    font-size: 36px;
+    color: #343a40;
+}
+
+.NecessityHeadBlock .day{
+    margin-top: 4px;
+    color: #868e96;
+    font-size: 21px;
+}
+
+.NecessityHeadBlock .tasks-left{
+    color: #20c997;
+    font-size: 18px;
+    margin-top: 40px;
+    font-weight: bold;
+}
