+++ orange/src/components/index.tsx
@@ -1,7 +1,7 @@
 import Header from './Header/Header';
 import Login from './Login/Login';
-import NecessityCreateModal from './Necessity/NecessityCreateModal';
-import NecessityHead from './Necessity/NecessityHead';
+import NecessityCreateModal from './Necessity/NecessityCreateModal/NecessityCreateModal';
+import NecessityHead from './Necessity/NecessityHead/NecessityHead';
 import NecessityItem from './Necessity/NecessityItem';
 import NecessityList from './Necessity/NecessityList';
 import NecessityTemplate from './Necessity/NecessityTemplate';
